Made the wind a part of the course
I prefer the semantics of this. The Course class already represents
the environment in which the disc is flying. It only makes sense that
the course also encodes the wind since that's part of the course.
This also opens up the possibility of per-hole wind conditions without
significant change to the main loop.